Update contract definitions
