+++ package.json
@@ -20,7 +20,8 @@
     "lint:fix": "eslint --fix . --ext .ts,.js",
     "verify": "npm run lint && npm run test",
     "sandbox": "DEBUG=ERROR-*,WARN-*,INFO-*,DEBUG-*,-DEBUG-helper*,-DEBUG-util* nodemon -r dotenv/config -r tsconfig-paths/register",
-    "contracts": "rimraf \"build/contracts\" && npx typechain --target=web3-v1 --outDir build/contracts \"src/contracts/*.json\" && prettier --write build/contracts/*.ts"
+    "contracts": "rimraf \"src/contracts/gen\" && npx typechain --target=web3-v1 --outDir src/contracts/gen \"src/contracts/abi/*.json\" && prettier --write src/contracts/gen/*.ts",
+    "contracts2": "rimraf \"build/contracts\" && npx typechain --target=web3-v1 --outDir build/contracts \"src/contracts/*.json\" && prettier --write build/contracts/*.ts"
   },
   "husky": {
     "hooks": {
@@ -68,6 +69,6 @@
     "typescript": "^3.7.2",
     "typescript-transform-paths": "^1.1.13",
     "typechain": "^1.0.3",
-    "typechain-target-web3-v1": "^1.0.0"
+    "typechain-target-web3-v1": "^1.0.4"
   }
 }
