+++ src/contracts/gen/BatchExchange.d.ts
@@ -1,0 +1,193 @@
+/* Generated by ts-generator ver. 0.0.8 */
+/* tslint:disable */
+
+import BN from 'bn.js'
+import { Contract, ContractOptions } from 'web3-eth-contract'
+import { EventLog } from 'web3-core'
+import { EventEmitter } from 'events'
+import { ContractEvent, Callback, TransactionObject, BlockType } from './types'
+
+interface EventOptions {
+  filter?: object
+  fromBlock?: BlockType
+  topics?: string[]
+}
+
+export class BatchExchange extends Contract {
+  constructor(jsonInterface: any[], address?: string, options?: ContractOptions)
+  clone(): BatchExchange
+  methods: {
+    getSecondsRemainingInBatch(): TransactionObject<string>
+
+    feeDenominator(): TransactionObject<string>
+
+    placeOrder(
+      buyToken: number | string,
+      sellToken: number | string,
+      validUntil: number | string,
+      buyAmount: number | string,
+      sellAmount: number | string,
+    ): TransactionObject<string>
+
+    submitSolution(
+      batchIndex: number | string,
+      claimedObjectiveValue: number | string,
+      owners: string[],
+      orderIds: (number | string)[],
+      buyVolumes: (number | string)[],
+      prices: (number | string)[],
+      tokenIdsForPrice: (number | string)[],
+    ): TransactionObject<string>
+
+    tokenIdToAddressMap(id: number | string): TransactionObject<string>
+
+    getPendingWithdrawAmount(user: string, token: string): TransactionObject<string>
+
+    requestWithdraw(token: string, amount: number | string): TransactionObject<void>
+
+    getPendingDepositAmount(user: string, token: string): TransactionObject<string>
+
+    deposit(token: string, amount: number | string): TransactionObject<void>
+
+    getPendingWithdrawBatchNumber(user: string, token: string): TransactionObject<string>
+
+    TOKEN_ADDITION_FEE_IN_OWL(): TransactionObject<string>
+
+    AMOUNT_MINIMUM(): TransactionObject<string>
+
+    feeToken(): TransactionObject<string>
+
+    placeValidFromOrders(
+      buyTokens: (number | string)[],
+      sellTokens: (number | string)[],
+      validFroms: (number | string)[],
+      validUntils: (number | string)[],
+      buyAmounts: (number | string)[],
+      sellAmounts: (number | string)[],
+    ): TransactionObject<string[]>
+
+    currentPrices(arg0: number | string): TransactionObject<string>
+
+    orders(
+      arg0: string,
+      arg1: number | string,
+    ): TransactionObject<{
+      buyToken: string
+      sellToken: string
+      validFrom: string
+      validUntil: string
+      priceNumerator: string
+      priceDenominator: string
+      usedAmount: string
+      0: string
+      1: string
+      2: string
+      3: string
+      4: string
+      5: string
+      6: string
+    }>
+
+    numTokens(): TransactionObject<string>
+
+    lastCreditBatchId(arg0: string, arg1: string): TransactionObject<string>
+
+    hasToken(addr: string): TransactionObject<boolean>
+
+    latestSolution(): TransactionObject<{
+      batchId: string
+      solutionSubmitter: string
+      feeReward: string
+      objectiveValue: string
+      0: string
+      1: string
+      2: string
+      3: string
+    }>
+
+    cancelOrders(ids: (number | string)[]): TransactionObject<void>
+
+    acceptingSolutions(batchIndex: number | string): TransactionObject<boolean>
+
+    getEncodedAuctionElements(): TransactionObject<string>
+
+    addToken(token: string): TransactionObject<void>
+
+    getBalance(user: string, token: string): TransactionObject<string>
+
+    freeStorageOfOrder(ids: (number | string)[]): TransactionObject<void>
+
+    BATCH_TIME(): TransactionObject<string>
+
+    getCurrentBatchId(): TransactionObject<string>
+
+    tokenAddressToIdMap(addr: string): TransactionObject<string>
+
+    requestFutureWithdraw(token: string, amount: number | string, batchId: number | string): TransactionObject<void>
+
+    hasValidWithdrawRequest(user: string, token: string): TransactionObject<boolean>
+
+    MAX_TOKENS(): TransactionObject<string>
+
+    getPendingDepositBatchNumber(user: string, token: string): TransactionObject<string>
+
+    withdraw(user: string, token: string): TransactionObject<void>
+
+    MAX_TOUCHED_ORDERS(): TransactionObject<string>
+
+    getCurrentObjectiveValue(): TransactionObject<string>
+  }
+  events: {
+    OrderPlacement: ContractEvent<{
+      owner: string
+      buyToken: string
+      sellToken: string
+      validFrom: string
+      validUntil: string
+      priceNumerator: string
+      priceDenominator: string
+      0: string
+      1: string
+      2: string
+      3: string
+      4: string
+      5: string
+      6: string
+    }>
+    OrderCancelation: ContractEvent<{
+      owner: string
+      id: string
+      0: string
+      1: string
+    }>
+    Deposit: ContractEvent<{
+      user: string
+      token: string
+      amount: string
+      stateIndex: string
+      0: string
+      1: string
+      2: string
+      3: string
+    }>
+    WithdrawRequest: ContractEvent<{
+      user: string
+      token: string
+      amount: string
+      stateIndex: string
+      0: string
+      1: string
+      2: string
+      3: string
+    }>
+    Withdraw: ContractEvent<{
+      user: string
+      token: string
+      amount: string
+      0: string
+      1: string
+      2: string
+    }>
+    allEvents: (options?: EventOptions, cb?: Callback<EventLog>) => EventEmitter
+  }
+}
