+++ src/index.ts
@@ -12,7 +12,7 @@ export * from './utils'
 // TODO: For now the contract definitions are declared in dex-js. But they will be defined in dex-contracts
 export * from './contracts/Erc20Contract'
 export * from './contracts/BatchExchangeContract'
-export * from './contracts/types'
+export * from './contracts/gen/types'
 
 // Json list
 export const tokenList: TokenDetailsConfig[] = require('tokenList.json')
