+++ test/sandbox/contracts/printEncodedFunctions.ts
@@ -10,8 +10,8 @@ require('dotenv').config()
  */
 const log = new Logger('sandbox:printEncodedFunctions')
 
-async function exec(): Promise<void> {
-  const abi = require('contracts/BatchExchange.json')
+async function exec (): Promise<void> {
+  const abi = require('contracts/abi/BatchExchange.json')
   const functions = abi.filter((def: AbiItem) => def.type === 'function')
 
   log.info('Found %d functions:', functions.length)
