+++ test/sandbox/contracts/printEventTopics.ts
@@ -11,7 +11,7 @@ require('dotenv').config()
 const log = new Logger('sandbox:contracts/printEventTopics')
 
 async function exec (): Promise<void> {
-  const abi = require('contracts/BatchExchangeAbi.json')
+  const abi = require('contracts/abi/BatchExchange.json')
   const events = abi.filter((def: AbiItem) => def.type === 'event')
 
   log.info('Found %d events:', events.length)
